v1.0.2

+++ commands/clear.js
@@ -6,6 +6,13 @@ async function clear(interaction, amount = 100) {
       return;
     }
 
+    amount = parseInt(amount);
+
+    if (isNaN(amount) || amount < 1 || amount > 100) {
+      await interaction.reply('Пожалуйста, выберите количество сообщений от 1 до 100.');
+      return;
+    }
+    
     const messages = await textChannel.messages.fetch({ limit: amount });
     await textChannel.bulkDelete(messages, true);
     await interaction.reply(`Удалено ${messages.size} сообщений.`);
